Refactor catalog.rs into a module
