+++ src/config/context.rs
@@ -17,7 +17,7 @@ use object_store::{local::LocalFileSystem, memory::InMemory, ObjectStore};
 #[cfg(feature = "catalog-postgres")]
 use crate::repository::postgres::PostgresRepository;
 
-use crate::catalog::{Error, Metastore};
+use crate::catalog::{metastore::Metastore, CatalogError};
 use crate::object_store::http::add_http_object_store;
 use crate::object_store::wrapped::InternalObjectStore;
 #[cfg(feature = "remote-tables")]
@@ -178,13 +178,13 @@ pub async fn build_context(cfg: &schema::SeafowlConfig) -> Result<SeafowlContext
     let metastore = build_metastore(cfg, internal_object_store.clone()).await;
 
     // Create default DB/collection
-    if let Err(Error::CatalogDoesNotExist { .. }) =
+    if let Err(CatalogError::CatalogDoesNotExist { .. }) =
         metastore.catalogs.get(DEFAULT_DB).await
     {
         metastore.catalogs.create(DEFAULT_DB).await.unwrap();
     }
 
-    if let Err(Error::SchemaDoesNotExist { .. }) =
+    if let Err(CatalogError::SchemaDoesNotExist { .. }) =
         metastore.schemas.get(DEFAULT_DB, DEFAULT_SCHEMA).await
     {
         metastore.schemas.create(DEFAULT_DB, DEFAULT_SCHEMA).await?;
